+++ src/App.tsx
@@ -1,0 +1,4 @@
+import React from 'react';
+import { AppRouter } from './router/app-router';
+
+export const App: React.FunctionComponent = () => <AppRouter />;
