--- src/app/app.tsx
@@ -1,15 +1,0 @@
-import { Main } from '../main/main';
-
-type AppProps = {
-  promoName: string;
-  promoGenre: string;
-  promoDate: string;
-};
-
-export const App: React.FunctionComponent<AppProps> = ({
-  promoName,
-  promoGenre,
-  promoDate,
-}: AppProps) => (
-  <Main promoName={promoName} promoGenre={promoGenre} promoDate={promoDate} />
-);
