+++ src/components/film-card/props.ts
@@ -1,0 +1,12 @@
+interface ImgProps {
+    src?: string;
+    alt?: string;
+    bgSrc?: string;
+  }
+
+export interface FilmProps {
+    img: ImgProps;
+    title: string;
+    genre: string;
+    year: number;
+  }
