+++ src/components/footer/footer.tsx
@@ -1,0 +1,12 @@
+import React from 'react';
+import { Logo } from '../logo/logo';
+
+export const Footer: React.FC = () => (
+  <footer className="page-footer">
+    <Logo className='logo__link--light' />
+
+    <div className="copyright">
+      <p>© 2019 What to watch Ltd.</p>
+    </div>
+  </footer>
+);
