+++ src/data/films-data.ts
@@ -1,0 +1,197 @@
+
+const DEFAULT_WIDTH = 280;
+const DEFAULT_HEIGHT = 175;
+
+export interface FilmProps {
+    id: number;
+    title: string;
+    imageSrc: string;
+    alt: string;
+    width: number;
+    height: number;
+    link: string;
+  }
+
+
+export const filmsData: FilmProps[] = [
+  {
+    id: 1,
+    title: 'Fantastic Beasts: The Crimes of Grindelwald',
+    imageSrc: 'img/fantastic-beasts-the-crimes-of-grindelwald.jpg',
+    alt: 'Fantastic Beasts: The Crimes of Grindelwald',
+    width: DEFAULT_WIDTH,
+    height: DEFAULT_HEIGHT,
+    link: 'film-page.html',
+  },
+  {
+    id: 2,
+    title: 'Bohemian Rhapsody',
+    imageSrc: 'img/bohemian-rhapsody.jpg',
+    alt: 'Bohemian Rhapsody',
+    width: DEFAULT_WIDTH,
+    height: DEFAULT_HEIGHT,
+    link: 'film-page.html',
+  },
+  {
+    id: 3,
+    title: 'Macbeth',
+    imageSrc: 'img/macbeth.jpg',
+    alt: 'Macbeth',
+    width: DEFAULT_WIDTH,
+    height: DEFAULT_HEIGHT,
+    link: 'film-page.html',
+  },
+  {
+    id: 4,
+    title: 'Aviator',
+    imageSrc: 'img/aviator.jpg',
+    alt: 'Aviator',
+    width: DEFAULT_WIDTH,
+    height: DEFAULT_HEIGHT,
+    link: 'film-page.html',
+  },
+  {
+    id: 5,
+    title: 'We need to talk about Kevin',
+    imageSrc: 'img/we-need-to-talk-about-kevin.jpg',
+    alt: 'We need to talk about Kevin',
+    width: DEFAULT_WIDTH,
+    height: DEFAULT_HEIGHT,
+    link: 'film-page.html',
+  },
+  {
+    id: 6,
+    title: 'What We Do in the Shadows',
+    imageSrc: 'img/what-we-do-in-the-shadows.jpg',
+    alt: 'What We Do in the Shadows',
+    width: DEFAULT_WIDTH,
+    height: DEFAULT_HEIGHT,
+    link: 'film-page.html',
+  },
+  {
+    id: 7,
+    title: 'Revenant',
+    imageSrc: 'img/revenant.jpg',
+    alt: 'Revenant',
+    width: DEFAULT_WIDTH,
+    height: DEFAULT_HEIGHT,
+    link: 'film-page.html',
+  },
+  {
+    id: 8,
+    title: 'Johnny English',
+    imageSrc: 'img/johnny-english.jpg',
+    alt: 'Johnny English',
+    width: DEFAULT_WIDTH,
+    height: DEFAULT_HEIGHT,
+    link: 'film-page.html',
+  },
+  {
+    id: 9,
+    title: 'Shutter Island',
+    imageSrc: 'img/shutter-island.jpg',
+    alt: 'Shutter Island',
+    width: DEFAULT_WIDTH,
+    height: DEFAULT_HEIGHT,
+    link: 'film-page.html',
+  },
+  {
+    id: 10,
+    title: 'Pulp Fiction',
+    imageSrc: 'img/pulp-fiction.jpg',
+    alt: 'Pulp Fiction',
+    width: DEFAULT_WIDTH,
+    height: DEFAULT_HEIGHT,
+    link: 'film-page.html',
+  },
+  {
+    id: 11,
+    title: 'No Country for Old Men',
+    imageSrc: 'img/no-country-for-old-men.jpg',
+    alt: 'No Country for Old Men',
+    width: DEFAULT_WIDTH,
+    height: DEFAULT_HEIGHT,
+    link: 'film-page.html',
+  },
+  {
+    id: 12,
+    title: 'Snatch',
+    imageSrc: 'img/snatch.jpg',
+    alt: 'Snatch',
+    width: DEFAULT_WIDTH,
+    height: DEFAULT_HEIGHT,
+    link: 'film-page.html',
+  },
+  {
+    id: 13,
+    title: 'Moonrise Kingdom',
+    imageSrc: 'img/moonrise-kingdom.jpg',
+    alt: 'Moonrise Kingdom',
+    width: DEFAULT_WIDTH,
+    height: DEFAULT_HEIGHT,
+    link: 'film-page.html',
+  },
+  {
+    id: 14,
+    title: 'Seven Years in Tibet',
+    imageSrc: 'img/seven-years-in-tibet.jpg',
+    alt: 'Seven Years in Tibet',
+    width: DEFAULT_WIDTH,
+    height: DEFAULT_HEIGHT,
+    link: 'film-page.html',
+  },
+  {
+    id: 15,
+    title: 'Midnight Special',
+    imageSrc: 'img/midnight-special.jpg',
+    alt: 'Midnight Special',
+    width: DEFAULT_WIDTH,
+    height: DEFAULT_HEIGHT,
+    link: 'film-page.html',
+  },
+  {
+    id: 16,
+    title: 'War of the Worlds',
+    imageSrc: 'img/war-of-the-worlds.jpg',
+    alt: 'War of the Worlds',
+    width: DEFAULT_WIDTH,
+    height: DEFAULT_HEIGHT,
+    link: 'film-page.html',
+  },
+  {
+    id: 17,
+    title: 'Dardjeeling Limited',
+    imageSrc: 'img/dardjeeling-limited.jpg',
+    alt: 'Dardjeeling Limited',
+    width: DEFAULT_WIDTH,
+    height: DEFAULT_HEIGHT,
+    link: 'film-page.html',
+  },
+  {
+    id: 18,
+    title: 'Orlando',
+    imageSrc: 'img/orlando.jpg',
+    alt: 'Orlando',
+    width: DEFAULT_WIDTH,
+    height: DEFAULT_HEIGHT,
+    link: 'film-page.html',
+  },
+  {
+    id: 19,
+    title: 'Mindhunter',
+    imageSrc: 'img/mindhunter.jpg',
+    alt: 'Mindhunter',
+    width: DEFAULT_WIDTH,
+    height: DEFAULT_HEIGHT,
+    link: 'film-page.html',
+  },
+  {
+    id: 20,
+    title: 'Midnight Special',
+    imageSrc: 'img/midnight-special.jpg',
+    alt: 'Midnight Special',
+    width: DEFAULT_WIDTH,
+    height: DEFAULT_HEIGHT,
+    link: 'film-page.html',
+  },
+];
