+++ src/pages/movie-page-details/movie-page-details.tsx
@@ -1,0 +1,225 @@
+import React from 'react';
+import { Link } from 'react-router-dom';
+import { Footer } from '../../components/footer/footer';
+
+export const MoviePageDetails: React.FunctionComponent = () => (
+  <>
+    <section className="film-card film-card--full">
+      <div className="film-card__hero">
+        <div className="film-card__bg">
+          <img
+            src="img/bg-the-grand-budapest-hotel.jpg"
+            alt="The Grand Budapest Hotel"
+          />
+        </div>
+
+        <h1 className="visually-hidden">WTW</h1>
+
+        <header className="page-header film-card__head">
+          <div className="logo">
+            <Link to="/" className="logo__link">
+              <span className="logo__letter logo__letter--1">W</span>
+              <span className="logo__letter logo__letter--2">T</span>
+              <span className="logo__letter logo__letter--3">W</span>
+            </Link>
+          </div>
+
+          <ul className="user-block">
+            <li className="user-block__item">
+              <div className="user-block__avatar">
+                <img
+                  src="img/avatar.jpg"
+                  alt="User avatar"
+                  width="63"
+                  height="63"
+                />
+              </div>
+            </li>
+            <li className="user-block__item">
+              <a className="user-block__link">Sign out</a>
+            </li>
+          </ul>
+        </header>
+
+        <div className="film-card__wrap">
+          <div className="film-card__desc">
+            <h2 className="film-card__title">The Grand Budapest Hotel</h2>
+            <p className="film-card__meta">
+              <span className="film-card__genre">Drama</span>
+              <span className="film-card__year">2014</span>
+            </p>
+
+            <div className="film-card__buttons">
+              <button className="btn btn--play film-card__button" type="button">
+                <svg viewBox="0 0 19 19" width="19" height="19">
+                  <use xlinkHref="#play-s"></use>
+                </svg>
+                <span>Play</span>
+              </button>
+              <button className="btn btn--list film-card__button" type="button">
+                <svg viewBox="0 0 19 20" width="19" height="20">
+                  <use xlinkHref="#add"></use>
+                </svg>
+                <span>My list</span>
+                <span className="film-card__count">9</span>
+              </button>
+              <Link to="add-review.html" className="btn film-card__button">
+                Add review
+              </Link>
+            </div>
+          </div>
+        </div>
+      </div>
+
+      <div className="film-card__wrap film-card__translate-top">
+        <div className="film-card__info">
+          <div className="film-card__poster film-card__poster--big">
+            <img
+              src="img/the-grand-budapest-hotel-poster.jpg"
+              alt="The Grand Budapest Hotel poster"
+              width="218"
+              height="327"
+            />
+          </div>
+
+          <div className="film-card__desc">
+            <nav className="film-nav film-card__nav">
+              <ul className="film-nav__list">
+                <li className="film-nav__item">
+                  <Link to="#" className="film-nav__link">
+                    Overview
+                  </Link>
+                </li>
+                <li className="film-nav__item film-nav__item--active">
+                  <Link to="#" className="film-nav__link">
+                    Details
+                  </Link>
+                </li>
+                <li className="film-nav__item">
+                  <Link to="#" className="film-nav__link">
+                    Reviews
+                  </Link>
+                </li>
+              </ul>
+            </nav>
+
+            <div className="film-card__text film-card__row">
+              <div className="film-card__text-col">
+                <p className="film-card__details-item">
+                  <strong className="film-card__details-name">Director</strong>
+                  <span className="film-card__details-value">Wes Anderson</span>
+                </p>
+                <p className="film-card__details-item">
+                  <strong className="film-card__details-name">Starring</strong>
+                  <span className="film-card__details-value">
+                    Bill Murray, <br />
+                    Edward Norton, <br />
+                    Jude Law, <br />
+                    Willem Dafoe, <br />
+                    Saoirse Ronan, <br />
+                    Tony Revoloru, <br />
+                    Tilda Swinton, <br />
+                    Tom Wilkinson, <br />
+                    Owen Wilkinson, <br />
+                    Adrien Brody, <br />
+                    Ralph Fiennes, <br />
+                    Jeff Goldblum
+                  </span>
+                </p>
+              </div>
+
+              <div className="film-card__text-col">
+                <p className="film-card__details-item">
+                  <strong className="film-card__details-name">Run Time</strong>
+                  <span className="film-card__details-value">1h 39m</span>
+                </p>
+                <p className="film-card__details-item">
+                  <strong className="film-card__details-name">Genre</strong>
+                  <span className="film-card__details-value">Comedy</span>
+                </p>
+                <p className="film-card__details-item">
+                  <strong className="film-card__details-name">Released</strong>
+                  <span className="film-card__details-value">2014</span>
+                </p>
+              </div>
+            </div>
+          </div>
+        </div>
+      </div>
+    </section>
+
+    <div className="page-content">
+      <section className="catalog catalog--like-this">
+        <h2 className="catalog__title">More like this</h2>
+
+        <div className="catalog__films-list">
+          <article className="small-film-card catalog__films-card">
+            <div className="small-film-card__image">
+              <img
+                src="img/fantastic-beasts-the-crimes-of-grindelwald.jpg"
+                alt="Fantastic Beasts: The Crimes of Grindelwald"
+                width="280"
+                height="175"
+              />
+            </div>
+            <h3 className="small-film-card__title">
+              <a className="small-film-card__link" href="film-page.html">
+                Fantastic Beasts: The Crimes of Grindelwald
+              </a>
+            </h3>
+          </article>
+
+          <article className="small-film-card catalog__films-card">
+            <div className="small-film-card__image">
+              <img
+                src="img/bohemian-rhapsody.jpg"
+                alt="Bohemian Rhapsody"
+                width="280"
+                height="175"
+              />
+            </div>
+            <h3 className="small-film-card__title">
+              <a className="small-film-card__link" href="film-page.html">
+                Bohemian Rhapsody
+              </a>
+            </h3>
+          </article>
+
+          <article className="small-film-card catalog__films-card">
+            <div className="small-film-card__image">
+              <img
+                src="img/macbeth.jpg"
+                alt="Macbeth"
+                width="280"
+                height="175"
+              />
+            </div>
+            <h3 className="small-film-card__title">
+              <a className="small-film-card__link" href="film-page.html">
+                Macbeth
+              </a>
+            </h3>
+          </article>
+
+          <article className="small-film-card catalog__films-card">
+            <div className="small-film-card__image">
+              <img
+                src="img/aviator.jpg"
+                alt="Aviator"
+                width="280"
+                height="175"
+              />
+            </div>
+            <h3 className="small-film-card__title">
+              <a className="small-film-card__link" href="film-page.html">
+                Aviator
+              </a>
+            </h3>
+          </article>
+        </div>
+      </section>
+
+      <Footer />
+    </div>
+  </>
+);
