+++ src/pages/player-pause/player-pause.tsx
