+++ src/router/consts.ts
@@ -1,0 +1,10 @@
+export const enum RouteLinks {
+    MAIN = '/',
+    LOGIN = '/login',
+    MY_LIST = '/mylist',
+    PLAYER = '/player/:id',
+    FILMS = '/films',
+    FILM = '/films/:id',
+    REVIEW = '/films/:id/review',
+    NOT_FOUND = '*',
+  }
